19-11-2021 bug fixes and patches

+++ src/graphql/mutations/vacancy.js
@@ -237,7 +237,7 @@ export const UPDATE_VACANCY_MUTATION = gql`
     $payRate: JobPayRate
     $minQualification: JobMinQualification
     $yearsOfExp: JobYearsOfExp
-    $salary: Int
+    $salary: Decimal
     $location: String
     $description: JSONString
     $positions: Int
